move morph_target_influence prop
